# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar PV Simulation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimulationInput {
  /**
   * Location latitude (-90 to 90)
   * @minimum -90
   * @maximum 90
   */
  latitude: number;
  /**
   * Location longitude (-180 to 180)
   * @minimum -180
   * @maximum 180
   */
  longitude: number;
  /**
   * Panel tilt angle in degrees (0 = horizontal, 90 = vertical)
   * @minimum 0
   * @maximum 90
   */
  tilt: number;
  /**
   * Panel azimuth angle in degrees (180 = south-facing)
   * @minimum 0
   * @maximum 360
   */
  azimuth: number;
  /**
   * Panel capacity in watts
   * @minimum 1
   */
  capacity_w: number;
  /** Simulation date (YYYY-MM-DD) */
  date: string;
  /** IANA timezone string (e.g. Asia/Seoul) */
  timezone?: string;
}

export interface SimulationResult {
  /** ISO8601 timestamps for each data point (5-min intervals) */
  timestamps: string[];
  /** AC power output in watts at each timestamp */
  ac_power_w: number[];
  /** Total daily energy production in Wh */
  total_energy_wh: number;
  /** Peak AC power output in watts */
  peak_power_w: number;
  /** Time of peak power output (ISO8601) */
  peak_time: string;
  /** Effective sunshine hours */
  sunshine_hours: number;
}

export interface OptimalTiltInput {
  /**
   * @minimum -90
   * @maximum 90
   */
  latitude: number;
  /**
   * @minimum -180
   * @maximum 180
   */
  longitude: number;
  /** @minimum 1 */
  capacity_w: number;
  /** Simulation date (YYYY-MM-DD) */
  date: string;
  timezone?: string;
  /**
   * @minimum 0
   * @maximum 360
   */
  azimuth?: number;
}

export interface TiltEnergy {
  tilt: number;
  energy_wh: number;
}

export interface OptimalTiltResult {
  /** Optimal tilt angle in degrees */
  optimal_tilt: number;
  /** Energy production at optimal tilt in Wh */
  best_energy_wh: number;
  tilt_energies: TiltEnergy[];
}

export interface ErrorResponse {
  error: string;
}