# lib/api-zod/src/generated/types/optimalTiltInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar PV Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface OptimalTiltInput {
  /**
   * @minimum -90
   * @maximum 90
   */
  latitude: number;
  /**
   * @minimum -180
   * @maximum 180
   */
  longitude: number;
  /** @minimum 1 */
  capacity_w: number;
  /** Simulation date (YYYY-MM-DD) */
  date: string;
  timezone?: string;
  /**
   * @minimum 0
   * @maximum 360
   */
  azimuth?: number;
}