# lib/api-zod/src/generated/types/optimalTiltResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar PV Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { TiltEnergy } from "./tiltEnergy";

export interface OptimalTiltResult {
  /** Optimal tilt angle in degrees */
  optimal_tilt: number;
  /** Energy production at optimal tilt in Wh */
  best_energy_wh: number;
  tilt_energies: TiltEnergy[];
}