# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar PV Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationInput {
  /**
   * Location latitude (-90 to 90)
   * @minimum -90
   * @maximum 90
   */
  latitude: number;
  /**
   * Location longitude (-180 to 180)
   * @minimum -180
   * @maximum 180
   */
  longitude: number;
  /**
   * Panel tilt angle in degrees (0 = horizontal, 90 = vertical)
   * @minimum 0
   * @maximum 90
   */
  tilt: number;
  /**
   * Panel azimuth angle in degrees (180 = south-facing)
   * @minimum 0
   * @maximum 360
   */
  azimuth: number;
  /**
   * Panel capacity in watts
   * @minimum 1
   */
  capacity_w: number;
  /** Simulation date (YYYY-MM-DD) */
  date: string;
  /** IANA timezone string (e.g. Asia/Seoul) */
  timezone?: string;
}