# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar PV Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationResult {
  /** ISO8601 timestamps for each data point (5-min intervals) */
  timestamps: string[];
  /** AC power output in watts at each timestamp */
  ac_power_w: number[];
  /** Total daily energy production in Wh */
  total_energy_wh: number;
  /** Peak AC power output in watts */
  peak_power_w: number;
  /** Time of peak power output (ISO8601) */
  peak_time: string;
  /** Effective sunshine hours */
  sunshine_hours: number;
}