# lib/api-zod/src/generated/types/tiltEnergy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar PV Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface TiltEnergy {
  tilt: number;
  energy_wh: number;
}